Migrate user data between networked machines
Copy user data (personal, not profile-specific) from "old" computer to "new".

Once the transfer is complete, a task will be created on the "new" computer to run after the first logon of the user to finish migrating the data to their new profile.

Requires the "old" profile to have the same name as the "new" profile.