Create ElevatePSScript-inflexible.ps1
A solution to a problem that doesn't really exist - checks the permissions of the user. If user is an administrator, it starts the supplied script.

If user is not currently an admin, it will open an administrative PowerShell instance (with UAC prompt for elevation) and then run the supplied script.